* fixed a bug which caused terrain loading failure on x86-64 (thanks to heiko) * fixed a bug which caused terrain heights to be clipped at 127 * added experimental terrain collision detection (EXTREMELY buggy and unreliable) * added preliminary FSAA (generic MSAA method) support - set r_ext_multisample_samples to a power of 2 (2 for 2x FSAA, 4 for 4x, etc.); note that currently the game will fail if the FSAA mode you set is not supported by your video card; numbers > 4 will most probably fail on recent nVidia cards, and CSAA is not yet supported by SDL

+++ code/qcommon/qfiles.h
@@ -813,7 +813,7 @@ typedef struct dterPatch_s {
 	short			baseZ;
 	unsigned short	shader;
 	short			lightmap;
-	long			dummy2[65];
+	int				dummy2[65];	// thanks to heiko_ for pointing out a x86-64 bug here
 	byte			heightmap[9][9];
 } dterPatch_t;
